requirements.txt
faster-whisper>=1.2.1
anthropic>=0.85.0
fastapi>=0.128.0
uvicorn>=0.40.0
streamlit>=1.52.2
neo4j>=5.19.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=24.1.0
python-multipart>=0.0.9
pymupdf>=1.24.0
python-docx>=1.1.0
